requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.12
PyJWT==2.10.1
requests==2.32.3
